some changes to Android
After some research, we've found the following

Regarding Android Audio HAL, using it would supply more
widespread use, since all other lower level drivers (TinyAlsa, etc)
plug in to it.  The disadvantage is that some lower level drivers may
offer interfaces, such as to DMA or a blocking API which either are
not in Audio HAL or only appear in some distributions or after some
version.

AAudio appears to be the future for performance Audio on Android.  It
is a minimalist interface with a standard "stream" like interface.
Avaliable since Android O.

OpenSLES is older, based on a standard, but as in all standards isn't
standard so you have to work with Android OpenSL ES rather than a
standard.  It offers some codec support and other bells and whistles.

AudioFlinger is also available as a modular server which offers some
more processing like things like mixing and resampling.  It has a
complicated API for tracks and performance, especially since over time
it's requirements change for getting the results you want. Since zc
does or will be able to do a lot of this, I've removed it from the
list.  It can be added later but for now, for prioritisation we have

1. Audio HAL:   low level, can be used to build an AudioFlinger
or AAudio competitor for example
2. AAudio:   small, simple to interface, the future.
3. OpenSL ES: for applying to older systems.