cgo cross package + more deadline work/testing
The callback functions can now be passed to another
package by Go by means of function pointers and
a little attention to the .h

Deadlines are ok for playback.  For capture, the
logic is that if the C API receives a sample after
it would have passed in real time with respect to the first exchange time,
then it is considered late.

If the C API does double buffering and regular sized
buffers, then this is correct.  Otherwise, it's just
a guess with respect to the time represented by the
first buffer size.